# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tribd",
    "description": "Tributary daemon API",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.1.1"
  },
  "paths": {
...
    "/api/v1/destinations/format": {
      "post": {
        "operationId": "format_drive",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Drive formatted; it remounts on its own"
          },
          "409": {
            "description": "Recording, or the drive is in use"
          },
          "422": {
            "description": "Refused — not a removable whole disk, or a bad label"
          },
          "501": {
            "description": "This installation has no format helper"
          }
        }
      }
    },
...
    "/api/v1/devices/profile": {
      "put": {
        "summary": "Switch a sound card's profile.",
        "description": "The active profile decides how many channels the card's devices expose,\nso this is how an interface with more inputs than the current profile\nshows gets the rest of them. The switch renames, resizes and remaps the\ncard's devices, so the whole document is re-enumerated and returned —\ncallers must replace their copy rather than patch it.",
        "operationId": "set_card_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCardProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Devices after the profile switch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceReport"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Recording in progress"
          },
          "422": {
            "description": "The card or profile was refused"
          }
        }
      }
    },
...
    "/api/v1/master": {
      "put": {
        "operationId": "update_master",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MasterPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The master section after the patch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MasterState"
                }
              }
            }
          },
          "422": {
            "description": "A value out of range"
          }
        }
      }
    },
    "/api/v1/sessions": {
      "get": {
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "description": "Sessions on the active destination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionsDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created and opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "409": {
            "description": "Recording, or that name is taken"
          },
          "422": {
            "description": "Bad name"
          }
        }
      }
    },
    "/api/v1/sessions/{id}": {
      "put": {
        "operationId": "rename_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session directory name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Renamed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "404": {
            "description": "No session by that id"
          },
          "422": {
            "description": "Bad name"
          }
        }
      },
      "delete": {
        "operationId": "delete_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session directory name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted, permanently"
          },
          "404": {
            "description": "No session by that id"
          },
          "409": {
            "description": "Recording, or that session is open"
          }
        }
      }
    },
    "/api/v1/sessions/{id}/open": {
      "post": {
        "operationId": "open_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session directory name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Opened",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          },
          "404": {
            "description": "No session by that id"
          },
          "409": {
            "description": "Recording"
          }
        }
      }
    },
    "/api/v1/settings/recording": {
      "get": {
        "operationId": "get_recording",
        "responses": {
          "200": {
            "description": "Current recording settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingSettingsDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "update_recording",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordingSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Settings after the update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingSettingsDto"
                }
              }
            }
          },
          "409": {
            "description": "Recording in progress"
          },
          "422": {
            "description": "Invalid destination, format, or sample rate"
          }
        }
      }
    },
...
    "/api/v1/strips": {
      "post": {
        "operationId": "create_strip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewStrip"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "The strip, freshly taped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StripState"
                }
              }
            }
          },
          "409": {
            "description": "The console is full"
          }
        }
      }
    },
    "/api/v1/strips/{id}": {
      "put": {
        "operationId": "update_strip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Strip id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StripPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The strip after the patch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StripState"
                }
              }
            }
          },
          "404": {
            "description": "No such strip"
          },
          "422": {
            "description": "A value out of range"
          }
        }
      },
      "delete": {
        "operationId": "delete_strip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Strip id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Strip removed"
          },
          "404": {
            "description": "No such strip"
          }
        }
      }
    },
...
    "/api/v1/takes/{take}": {
      "delete": {
        "summary": "Permanently remove one take — audio, peaks and manifest together.",
        "description": "Refused while recording, and refused for the take being played: the\nfeeder threads hold open readers and reopen the files on every loop\npass, so unlinking underneath them would run playback off unlinked\ninodes until a later wrap died somewhere confusing.",
        "operationId": "delete_take",
        "parameters": [
          {
            "name": "take",
            "in": "path",
            "description": "Take number",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted; the remaining takes, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TakeDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No such take"
          },
          "409": {
            "description": "Recording, or that take is playing"
          }
        }
      }
    },
    "/api/v1/takes/{take}/peaks": {
      "get": {
        "summary": "Waveform peaks for every track of a take, in `TakeDto.tracks` order.",
        "description": "Binary layout (little-endian): `\"TPKS\"` · `version: u8 = 1` · `pad: u8`\n· `track_count: u16` · `samples_per_bin: u32` · then per track:\n`bin_count: u32` + `bin_count × (min: i16, max: i16)`.",
        "operationId": "take_peaks",
        "parameters": [
          {
            "name": "take",
            "in": "path",
            "description": "Take number",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Binary peaks document",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "No such take"
          }
        }
      }
    },
...
    "/api/v1/transport/lanes/{index}": {
      "put": {
        "operationId": "set_lane",
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "description": "Lane index within the latest take's tracks",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaneGateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Lane gates updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportDto"
                }
              }
            }
          },
          "422": {
            "description": "No such lane"
          }
        }
      }
    },
    "/api/v1/transport/loop": {
      "put": {
        "operationId": "set_loop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoopRegionDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Loop set (rebuilds a playing session)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportDto"
                }
              }
            }
          },
          "409": {
            "description": "Recording"
          },
          "422": {
            "description": "Region the take can't honor"
          }
        }
      },
      "delete": {
        "operationId": "clear_loop",
        "responses": {
          "200": {
            "description": "Loop cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transport/monitor": {
      "put": {
        "operationId": "set_monitor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Monitor target set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transport/record/start": {
      "post": {
        "operationId": "record_start",
        "responses": {
          "200": {
            "description": "Recording",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportDto"
                }
              }
            }
          },
          "409": {
            "description": "Already recording"
          },
          "422": {
            "description": "Nothing armed"
          }
        }
      }
    },
...
    "/api/v1/transport/take": {
      "put": {
        "summary": "Point the transport at a take of the open session.",
        "description": "A property of the transport, so PUT with a body — the same shape as\n`PUT /transport/loop` and `/monitor`, not a verb path like\n`record/start`. Selecting is what makes `lanes`, `total_frames` and the\nwaveform describe that take, which is why it is separate from PLAY: the\npoint is to be able to LOOK at a take.",
        "operationId": "select_take",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectTakeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Selected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportDto"
                }
              }
            }
          },
          "404": {
            "description": "No such take in the open session"
          },
          "409": {
            "description": "Recording"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceReport": {
        "type": "object",
        "description": "One row of the patchbay device document.",
        "required": [
          "name",
          "channels",
          "active",
          "status",
          "patched",
          "underruns",
          "overruns",
          "profiles",
          "muted"
        ],
        "properties": {
          "active": {
            "type": "boolean",
            "description": "The OS default input — what `device: null` patches feed from."
          },
          "card": {
            "type": [
              "string",
              "null"
            ],
            "description": "The card this device belongs to. A profile change names the CARD."
          },
          "channel_map": {
            "type": [
              "string",
              "null"
            ],
            "description": "The source's channel positions (\"aux0,aux1,…\"). NOT decoration:\nthe capture opens in exactly these positions, because the server\nroutes by position name and will synthesise a surround map that\ndrops most of a pro-audio source's inputs otherwise."
          },
          "channels": {
            "type": "integer",
            "format": "int32",
            "description": "What the device exposes RIGHT NOW — a property of the card's active\nprofile, not of the hardware. 0 = unknown (an absent device was\nnever enumerated this boot).",
            "minimum": 0
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Why capture isn't running, when the failure happened at open time\n(post-open stream deaths land in the journal only)."
          },
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "The friendly print the system sound menu shows, when the source\nlayer provides one."
          },
          "muted": {
            "type": "boolean",
            "description": "The source layer muted this input. It still opens, still streams,\nand every meter fed from it reads silence — so a dead meter has an\nexplanation here that exists nowhere else in the report."
          },
          "name": {
            "type": "string"
          },
          "overruns": {
            "type": "integer",
            "format": "int64",
            "description": "Whole frames dropped because the engine wasn't draining fast enough.",
            "minimum": 0
          },
          "patched": {
            "type": "boolean",
            "description": "Some strip references it (directly, via the default, or via an alias)."
          },
          "profile": {
            "type": [
              "string",
              "null"
            ],
            "description": "The card's active profile, when it has one."
          },
          "profiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileReport"
            },
            "description": "Profiles this device's card can be switched into. More than one\nmeans the channel count above is a choice, not a hardware limit."
          },
          "reconciled_from": {
            "type": [
              "string",
              "null"
            ],
            "description": "The stored project name this device was matched from, when name\nreconciliation adopted it under a changed name."
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          },
          "underruns": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "volume_percent": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The quietest channel's volume as a percentage of unity, when the\nsource layer says. None = unknown (raw ALSA, or an absent device),\nwhich is deliberately distinct from 0.",
            "minimum": 0
          }
        }
      },
...
      "FormatRequest": {
        "type": "object",
        "required": [
          "device",
          "label"
        ],
        "properties": {
          "device": {
            "type": "string",
            "description": "The whole disk, as `DriveDto.disk` names it (\"/dev/sda\"). A\npartition is refused: formatting one strands the rest of the drive."
          },
          "label": {
            "type": "string",
            "description": "The exFAT volume label, which becomes the drive's printed name."
          }
        }
      },
...
      "MasterPatch": {
        "type": "object",
        "properties": {
          "fader_db": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "MasterState": {
        "type": "object",
        "description": "The master output section: one stereo fader and its own record arm.",
        "required": [
          "fader_db",
          "record_arm"
        ],
        "properties": {
          "fader_db": {
            "type": "number",
            "format": "float"
          },
          "record_arm": {
            "type": "boolean"
          }
        }
      },
...
      "MonitorBody": {
        "type": "object",
        "required": [
          "target"
        ],
        "properties": {
          "target": {
            "$ref": "#/components/schemas/MonitorTarget"
          }
        }
      },
...
      "NewSession": {
        "type": "object",
        "required": [
          "name",
          "seed"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "seed": {
            "$ref": "#/components/schemas/SessionSeed",
            "description": "What the new session's desk starts from."
          }
        },
        "additionalProperties": false
      },
      "NewStrip": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "RecordingSettingsDto": {
        "type": "object",
        "required": [
          "destination",
          "default_destination",
          "project_name",
          "format",
          "active_sample_rate",
          "configured_sample_rate",
          "restart_required"
        ],
        "properties": {
          "active_sample_rate": {
            "type": "integer",
            "format": "int32",
            "description": "What the engine graph runs at — immutable per boot.",
            "minimum": 0
          },
          "configured_destination": {
            "type": [
              "string",
              "null"
            ],
            "description": "The persisted preference — differs from `destination` when the\nconfigured drive was missing at boot and the daemon fell back."
          },
          "configured_sample_rate": {
            "type": "integer",
            "format": "int32",
            "description": "What the prefs ask for — takes effect at the next daemon start.",
            "minimum": 0
          },
          "default_destination": {
            "type": "string",
            "description": "The boot-config fallback root."
          },
          "destination": {
            "type": "string",
            "description": "The projects root the next take lands on."
          },
          "format": {
            "$ref": "#/components/schemas/RecordFormat"
          },
          "project_name": {
            "type": "string",
            "description": "The open project at the destination."
          },
          "restart_required": {
            "type": "boolean"
          }
        }
      },
      "RenameSession": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "SessionDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at_unix",
          "take_count",
          "open"
        ],
        "properties": {
          "created_at_unix": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "description": "The directory name. Stable across a rename — renaming retitles the\nmanifest and never moves the directory, so an id a client is\nholding stays valid."
          },
          "name": {
            "type": "string"
          },
          "open": {
            "type": "boolean",
            "description": "Exactly one session in a list is open."
          },
          "take_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "SessionsDto": {
        "type": "object",
        "required": [
          "root",
          "root_present",
          "sessions"
        ],
        "properties": {
          "root": {
            "type": "string",
            "description": "The projects root these live under — the active destination."
          },
          "root_present": {
            "type": "boolean",
            "description": "False when the destination has gone away (drive unplugged). An\nempty list then means \"we cannot see them\", not \"there are none\" —\ntwo different sentences the console must not merge."
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionDto"
            },
            "description": "Newest first, in the same order an adopt would pick from."
          }
        }
      },
      "SetCardProfile": {
        "type": "object",
        "description": "Which profile to put a sound card into.",
        "required": [
          "card",
          "profile"
        ],
        "properties": {
          "card": {
            "type": "string",
            "description": "The card, as `DeviceReport.card` names it. Profiles belong to the\ncard, not to any one of its devices."
          },
          "profile": {
            "type": "string",
            "description": "One of that card's `DeviceReport.profiles[].name`."
          }
        }
      },
...
    }
  }
}